+++ src/briefy/choreographer/actions/slack/leica/professional.py
@@ -30,7 +30,7 @@ class ProfessionalSlack(Slack):
         payload['data'] = {
             'fields': [
                 {
-                    'title': 'Fullname',
+                    'title': 'Full name',
                     'value': data['title'],
                     'short': True,
                 },
@@ -100,7 +100,7 @@ class ProfessionalWfApprove(ProfessionalSlack):
         payload['data'] = {
             'fields': [
                 {
-                    'title': 'Fullname',
+                    'title': 'Full name',
                     'value': data['title'],
                     'short': True,
                 },
@@ -129,11 +129,54 @@ class ProfessionalWfApprove(ProfessionalSlack):
         return payload
 
 
+@adapter(events.IProfessionalWfApprove)
+@implementer(ISlack)
+class ProfessionalWfApproveFinance(ProfessionalSlack):
+    """After quality approving a Creative, post on Slack."""
+
+    _channel = '#creatives-validation'
+    title = 'A new creative was approved, please proceed with legal validation'
+
+    def transform(self) -> dict:
+        """Transform data."""
+        payload = super().transform()
+        data = self.data
+        payload['data'] = {
+            'fields': [
+                {
+                    'title': 'Full name',
+                    'value': data['title'],
+                    'short': True,
+                },
+                {
+                    'title': 'Email',
+                    'value': data['email'],
+                    'short': True,
+                },
+                {
+                    'title': 'Phone',
+                    'value': data['mobile'],
+                    'short': True,
+                }
+            ]
+        }
+        return payload
+
+
 @adapter(events.IProfessionalWfValidate)
 @implementer(ISlack)
 class ProfessionalWfValidate(ProfessionalSlack):
-    """After activating a new Professional, post on Slack."""
+    """After validating a new Professional, post on Slack."""
 
+    title = 'Creative was validated by legal'
+
+
+@adapter(events.IProfessionalWfValidate)
+@implementer(ISlack)
+class ProfessionalWfValidateScout(ProfessionalSlack):
+    """After validating a new Professional, post on Slack."""
+
+    _channel = '#scouting-team'
     title = 'Creative was validated by legal'